Refactor move pydantic model to datatypes/_schema.py
