+++ lib/galaxy/webapps/galaxy/api/datatypes.py
@@ -23,7 +23,9 @@ from galaxy.managers.datatypes import (
     DatatypeDetails,
     DatatypesCombinedMap,
     DatatypesMap,
-    view_converters, view_edam_data, view_edam_formats,
+    view_converters,
+    view_edam_data,
+    view_edam_formats,
     view_index,
     view_mapping,
     view_sniffers,
